chore: Rename components
