+++ src/routes/view/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import View from '$lib/components/view.svelte';
+  import View from '$lib/components/View.svelte';
   import { initHandler } from '$lib/util/util';
   import { onMount } from 'svelte';
   onMount(initHandler);
